+++ todoman/model.py
@@ -543,8 +543,10 @@ class Cache:
 
         return rv
 
+
     def todos(self, all=False, lists=[], priority=None, location='',
-              category='', grep='', sort=[], reverse=True, due=None):
+              category='', grep='', sort=[], reverse=True, due=None,
+              complete=None, start=None):
         """
         Returns filtered cached todos, in a specified order.
 
@@ -568,6 +570,9 @@ class Cache:
         :param int due: Return only todos due within ``due`` hours.
         :param str priority: Only return todos with priority at least as
             high as specified.
+        :param bool complete: If true, return completed tasks,
+            else incomplete tasks
+        :param start: Return only todos before/after ``start`` date
         :return: A sorted, filtered list of todos.
         :rtype: generator
         """
@@ -578,9 +583,12 @@ class Cache:
 
         if not all:
             # XXX: Duplicated logic of Todo.is_completed
-            extra_where.append('AND completed_at is NULL '
-                               'AND status != "CANCELLED" '
-                               'AND status != "COMPLETED"')
+            if complete:
+                extra_where.append('AND status == "COMPLETED"')
+            else:
+                extra_where.append('AND completed_at is NULL '
+                                   'AND status != "CANCELLED" '
+                                   'AND status != "COMPLETED"')
         if lists:
             lists = [l.name if isinstance(l, List) else l for l in lists]
             q = ', '.join(['?'] * len(lists))
@@ -605,7 +613,14 @@ class Cache:
             max_due = (datetime.now() + timedelta(hours=due)).timestamp()
             extra_where.append('AND due IS NOT NULL AND due < ?')
             params.append(max_due)
-
+        if start:
+            is_before, dt = start
+            if is_before:
+                extra_where.append('AND created_at <= ?')
+                params.append(dt)
+            else:
+                extra_where.append('AND created_at >= ?')
+                params.append(dt)
         if sort:
             order = []
             for s in sort: