+++ CHANGELOG.rst
@@ -14,6 +14,9 @@ v3.0.0
   parameter.
 * Use the system's date format as a default.
 * Show "Today" or "Tomorrow" when due date is today/tomorrow respectively.
+* Add ``--start="[before|after] <date>`` option for ``todo list``. It shows todos
+  created before/after given date.
+* Add flag "--done-only" to todo list. Displays only completed tasks.
 
 v2.1.0
 ------